requirements.txt
Django==2.2.2
django-mdeditor==0.1.14
django-taggit==1.1.0
Markdown==3.1.1
Pygments==2.4.2
pytz==2019.1
sqlparse==0.3.0
